Add a cast to remove a warning

+++ source/arch/sys_arch.c
@@ -46,7 +46,7 @@ void SysTick_Init(void) {
     if (SysTick_Config(SystemCoreClock / 1000)) {
         while (1);     /* Capture error */
     }
-	NVIC_SetVector(SysTick_IRQn,&SysTick_Handler);
+	NVIC_SetVector(SysTick_IRQn,(uint32_t)&SysTick_Handler);
 }
 
 /** \brief  SysTick IRQ handler and timebase management
